+++ hackathon/ITK-V3D-Plugins/Source/Plugin2Call/pipeline/ITKShapeDetection_called.cxx
@@ -139,7 +139,7 @@ public:
         }
 
         int nx=this->m_NumberOfPixelsAlongX,ny=this->m_NumberOfPixelsAlongY,nz=this->m_NumberOfPixelsAlongZ;
-        const double stoppingTime = sqrt(nx*nx + ny*ny + nz*nz);
+        const double stoppingTime = sqrt(double(nx*nx + ny*ny + nz*nz));
